+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.2] - 2023-03-24
+
+### Added
+
+- `remove` action
+- `remove` test
+
+### Change
+
+- `refetch` to `execute`
+
+### Fix
+
+- `immediate` not use debounce timeout
+
 ## 0.0.1 - 2023-03-21
 
 ### Added
@@ -9,3 +24,5 @@ All notable changes to this project will be documented in this file.
 - `defineStoreRepository` a function to create a store repository;
 - `read` and `submit` actions;
 - `getQueryByName` and `getItemByKey` getters.
+
+[0.0.2]: https://github.com/volverjs/query-vue/compare/v0.0.1...v0.0.2